Keybindings Hook
- updated the gamepad preferences UI so it fits in the modal
- removed keybindings from default store
- changed the shuttleControlEvents so they now hold all the information needed for creating a keybind/shortcut
- added new hook called useKeybinding that adds the shuttleControlEvents for each widget to the store, and removes them from the store if they no longer exist
        - to help with keybind removal, added a new file called shuttleEvents.js that collects the shuttleControlEvents from every widget for the session (not persistent)
