+++ src/app/containers/Preferences/General/defaultMachineProfiles.js
@@ -50,6 +50,7 @@ export default [
         'spindle': false,
         'coolant': false,
         'laser': false,
+        'laserOnOutline': false,
         'eepromSettings': millOneFile,
     },
     {
@@ -72,6 +73,7 @@ export default [
         'spindle': false,
         'coolant': false,
         'laser': false,
+        'laserOnOutline': false,
         'eepromSettings': millOneFile,
     },
     {
@@ -94,6 +96,7 @@ export default [
         'spindle': false,
         'coolant': false,
         'laser': false,
+        'laserOnOutline': false,
         'eepromSettings': millOneV3File,
     },
     {
@@ -116,6 +119,7 @@ export default [
         'spindle': false,
         'coolant': false,
         'laser': false,
+        'laserOnOutline': false,
         'eepromSettings': longMill12x12File,
     },
     {
@@ -138,6 +142,7 @@ export default [
         'spindle': false,
         'coolant': false,
         'laser': false,
+        'laserOnOutline': false,
         'eepromSettings': longMill12x30File,
     },
     {
@@ -160,6 +165,7 @@ export default [
         'spindle': false,
         'coolant': false,
         'laser': false,
+        'laserOnOutline': false,
         'eepromSettings': longMill30x30File,
     },
     {
@@ -182,6 +188,7 @@ export default [
         'spindle': false,
         'coolant': false,
         'laser': false,
+        'laserOnOutline': false,
         'eepromSettings': MK2LongMill30x30File,
     },
     {
@@ -204,6 +211,7 @@ export default [
         'spindle': false,
         'coolant': false,
         'laser': false,
+        'laserOnOutline': false,
         'eepromSettings': MK2LongMill12x30File,
     },
     {
@@ -244,7 +252,8 @@ export default [
         'endstops': true,
         'spindle': true,
         'coolant': false,
-        'laser': false
+        'laser': false,
+        'laserOnOutline': false,
     },
     {
         'id': 19,
@@ -264,7 +273,8 @@ export default [
         'endstops': true,
         'spindle': true,
         'coolant': false,
-        'laser': false
+        'laser': false,
+        'laserOnOutline': false,
     },
     {
         'id': 22,
@@ -284,7 +294,8 @@ export default [
         'endstops': true,
         'spindle': false,
         'coolant': false,
-        'laser': false
+        'laser': false,
+        'laserOnOutline': false,
     },
     {
         'id': 23,
@@ -304,7 +315,8 @@ export default [
         'endstops': true,
         'spindle': false,
         'coolant': false,
-        'laser': false
+        'laser': false,
+        'laserOnOutline': false,
     },
     {
         'id': 33,
@@ -324,7 +336,8 @@ export default [
         'endstops': true,
         'spindle': false,
         'coolant': false,
-        'laser': false
+        'laser': false,
+        'laserOnOutline': false,
     },
     {
         'id': 34,
@@ -344,7 +357,8 @@ export default [
         'endstops': false,
         'spindle': false,
         'coolant': false,
-        'laser': false
+        'laser': false,
+        'laserOnOutline': false,
     },
     {
         'id': 41,
@@ -364,7 +378,8 @@ export default [
         'endstops': false,
         'spindle': false,
         'coolant': false,
-        'laser': false
+        'laser': false,
+        'laserOnOutline': false,
     },
     {
         'id': 43,
@@ -384,7 +399,8 @@ export default [
         'endstops': true,
         'spindle': false,
         'coolant': false,
-        'laser': false
+        'laser': false,
+        'laserOnOutline': false,
     },
     {
         'id': 44,
@@ -404,7 +420,8 @@ export default [
         'endstops': true,
         'spindle': true,
         'coolant': false,
-        'laser': true
+        'laser': false,
+        'laserOnOutline': false,
     },
     {
         'id': 45,
@@ -424,7 +441,8 @@ export default [
         'endstops': false,
         'spindle': false,
         'coolant': false,
-        'laser': false
+        'laser': false,
+        'laserOnOutline': false,
     },
 
     {
@@ -445,7 +463,8 @@ export default [
         'endstops': false,
         'spindle': false,
         'coolant': false,
-        'laser': false
+        'laser': false,
+        'laserOnOutline': false,
     },
     {
         'id': 50,
@@ -465,7 +484,8 @@ export default [
         'endstops': false,
         'spindle': false,
         'coolant': false,
-        'laser': false
+        'laser': false,
+        'laserOnOutline': false,
     },
     {
         'id': 51,
@@ -485,7 +505,8 @@ export default [
         'endstops': false,
         'spindle': false,
         'coolant': false,
-        'laser': false
+        'laser': false,
+        'laserOnOutline': false,
     },
     {
         'id': 52,
@@ -505,6 +526,7 @@ export default [
         'endstops': false,
         'spindle': false,
         'coolant': false,
-        'laser': false
+        'laser': false,
+        'laserOnOutline': false,
     },
 ];
