+++ src/app/containers/Preferences/Keybindings/index.styl
@@ -4,10 +4,16 @@ $primary = #3e85c7;
   display: grid;
   grid-template-rows: 19fr 1fr;
   margin-top: 0.5rem;
-  height: 565px;
+  max-height: 565px;
 }
 .table-wrapper {
   height: 528px;
+  overflow-y: scroll;
+  background-color: white;
+  overflow-x: hidden;
+}
+.table-wrapper-controller {
+  max-height: 445px;
   overflow-y: scroll;
   background-color: white;
   overflow-x: hidden;
