+++ src/app/containers/Preferences/Preferences.jsx
@@ -159,7 +159,7 @@ class PreferencesPage extends PureComponent {
                 objects: this.visualizerConfig.get('objects'),
                 disabled: this.visualizerConfig.get('disabled'),
                 disabledLite: this.visualizerConfig.get('disabledLite'),
-                showSoftLimitsWarning: this.visualizerConfig.get('showSoftLimitsWarning')
+                showSoftLimitsWarning: this.visualizerConfig.get('showSoftLimitsWarning', false)
             },
             showWarning: store.get('widgets.visualizer.showWarning'),
             showLineWarnings: store.get('widgets.visualizer.showLineWarnings'),
