+++ src/app/index.jsx
@@ -60,6 +60,7 @@ import Modal from './components/Modal';
 import Space from './components/Space';
 import './styles/vendor.styl';
 import './styles/app.styl';
+import { removeOldKeybindings } from './lib/useKeybinding';
 
 const renderPage = () => {
     const container = document.createElement('div');
@@ -235,6 +236,8 @@ series([
     }
 
     renderPage();
+
+    removeOldKeybindings();
 }).catch(err => {
     log.error(err);
 });
