+++ src/app/lib/gamepad/index.js
@@ -1,6 +1,7 @@
 import { GamepadListener } from 'gamepad.js';
 import store from 'app/store';
 import { Toaster, TOASTER_INFO } from 'app/lib/toaster/ToasterLib';
+import shuttleEvents from '../shuttleEvents';
 
 const STOP_JOG_CMD = 'STOP_JOG';
 
@@ -110,8 +111,8 @@ export const runAction = ({ event, shuttleControlEvents }) => {
         return;
     }
 
-    const runEvent = shuttleControlEvents[action.cmd];
-
+    const allShuttleControlEvents = shuttleEvents.allShuttleControlEvents;
+    const runEvent = allShuttleControlEvents.find(element => element.cmd === action.cmd).callback;
     if (runEvent) {
         runEvent(null, action.payload);
     }
