+++ src/app/widgets/Coolant/CoolantControls.jsx
@@ -28,8 +28,9 @@ import FunctionButton from 'app/components/FunctionButton/FunctionButton';
 import controller from 'app/lib/controller';
 import combokeys from 'app/lib/combokeys';
 import gamepad, { runAction } from 'app/lib/gamepad';
-import { GRBL_ACTIVE_STATE_IDLE } from 'app/constants';
+import { GRBL_ACTIVE_STATE_IDLE, COOLANT_CATEGORY } from 'app/constants';
 import styles from './index.styl';
+import useKeybinding from '../../lib/useKeybinding';
 
 
 const sendM7 = () => {
@@ -43,14 +44,38 @@ const sendM9 = () => {
 };
 
 const shuttleControlEvents = {
-    MIST_COOLANT: () => {
-        sendM7();
+    MIST_COOLANT: {
+        title: 'Mist Coolant',
+        keys: '',
+        cmd: 'MIST_COOLANT',
+        preventDefault: false,
+        isActive: true,
+        category: COOLANT_CATEGORY,
+        callback: () => {
+            sendM7();
+        },
     },
-    FLOOD_COOLANT: () => {
-        sendM8();
+    FLOOD_COOLANT: {
+        title: 'Flood Coolant',
+        keys: '',
+        cmd: 'FLOOD_COOLANT',
+        preventDefault: false,
+        isActive: true,
+        category: COOLANT_CATEGORY,
+        callback: () => {
+            sendM8();
+        },
     },
-    STOP_COOLANT: () => {
-        sendM9();
+    STOP_COOLANT: {
+        title: 'Stop Coolant',
+        keys: '',
+        cmd: 'STOP_COOLANT',
+        preventDefault: false,
+        isActive: true,
+        category: COOLANT_CATEGORY,
+        callback: () => {
+            sendM9();
+        }
     }
 };
 
@@ -58,16 +83,14 @@ const subscribeShuttleControl = () => {
     combokeys.reload();
 
     Object.keys(shuttleControlEvents).forEach(eventName => {
-        const callback = shuttleControlEvents[eventName];
+        const callback = shuttleControlEvents[eventName].callback;
         combokeys.on(eventName, callback);
     });
-
-    gamepad.on('gamepad:button', (event) => runAction({ event, shuttleControlEvents: shuttleControlEvents }));
 };
 
 const unsubscribeShuttleControl = () => {
     Object.keys(shuttleControlEvents).forEach(eventName => {
-        const callback = shuttleControlEvents[eventName];
+        const callback = shuttleControlEvents[eventName].callback;
         combokeys.removeListener(eventName, callback);
     });
 };
@@ -75,6 +98,8 @@ const unsubscribeShuttleControl = () => {
 const CoolantControls = ({ canClick }) => {
     useEffect(() => {
         subscribeShuttleControl();
+        useKeybinding(shuttleControlEvents);
+        gamepad.on('gamepad:button', (event) => runAction({ event, shuttleControlEvents: shuttleControlEvents }));
         return function cleanup() {
             unsubscribeShuttleControl();
         };
